requirements.txt
pymongo
streamlit
openai
groq
python-dotenv
fpdf